+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 4.0.0
+
+- queryAsync now returns the whole result object instead of just the rows; queryRowsAsync remains unchanged
+  and should be used when only the result rows are wanted
+
+
 ## 3.2.1
 
 - Update dependencies